Testing update display after operator trigger.
I think this app needs a new action creator,
something like UPDATE_DISPLAY for cases like
when an user press an operand, then display it's updated,
and when hits an operator, display its also updated but
operands and operators states (global states) are too.

So, display updates actions are semi-independent from operands
and operators actions.

A somewhat similar case is displayed on this code.